Refactor, re-org translate panel-related components
