+++ src/javascript/ContentEditor/actions/contenteditor/translate/TranslatePanel/index.js
@@ -1,0 +1,1 @@
+export {TranslatePanel} from './TranslatePanel';
